Don't automock from pytest plugin setup, use fixture
We need to be able to decide which tests need to be automocked on the
project using the library.

Calling automock from the pytest plugin setup means every test will get
automocked and there is no way to disable this behaviour.

A fixture is more flexible, since it can be invoked from any test that
requires it and it can be injected to every test in the same
module/package using an `autouse=True` fixture if necessary.